requirements.txt
requests==2.24.0
colorama
termcolor
discord.py
python-dotenv